Proper initialization.

+++ include/core/constants.h
@@ -4,5 +4,5 @@
 namespace Black
 {
 	// 'Undefined' state of any index.
-	constexpr size_t UNDEFINED_INDEX = -1;
+	constexpr size_t UNDEFINED_INDEX = static_cast<size_t>( -1 );
 }
